Rename Actions to EntitiesActions
The module should reflect the components name.
